--- packages/ui/components/AppLayout/AppLayout.module.css
@@ -1,5 +1,0 @@
-.logo {
-  height: 32px;
-  margin: 16px;
-  background: rgba(255, 255, 255, 0.3);
-}
