--- packages/ui/components/DropDown/index.tsx
@@ -1,1 +1,0 @@
-export { default as DropDown } from './DropDown'
