--- packages/ui/components/Editor/editor.module.css
@@ -1,8 +1,0 @@
-.editor {
-    margin-top: 120px;
-
-}
-
-.ck-editor__editable_inline {
-    min-height: 500px;
-}
