--- packages/ui/components/Editor/index.tsx
@@ -1,4 +1,0 @@
-import { EncodeIntoResult } from 'util'
-import Editor from './Editor'
-
-export { Editor as default }
