+++ packages/ui/pages/programs/[[...id]].tsx
@@ -3,7 +3,7 @@ import { useQuery } from 'react-query'
 import axios from 'axios'
 import { Button, Result, Skeleton, Space } from 'antd'
 import Link from 'next/link'
-import { AppLayout } from 'components/AppLayout'
+import AppLayout from 'components/AppLayout'
 import styled from 'styled-components'
 import { PlusOutlined } from '@ant-design/icons'
 import OTTable from 'components/TasksTable'
